Use annotations instead of Velocity to specify different behavior for different API levels.
Use Build.VERSION_CODE values instead of integers in @Implementation/@Implements and conditionals.

+++ robolectric/src/test/java/org/robolectric/shadows/ShadowSQLiteConnectionTest.java
@@ -45,7 +45,7 @@ public class ShadowSQLiteConnectionTest {
     SQLiteStatement createStatement = database.compileStatement(
         "CREATE TABLE `routine` (`id` INTEGER PRIMARY KEY AUTOINCREMENT , `name` VARCHAR , `lastUsed` INTEGER DEFAULT 0 ,  UNIQUE (`name`)) ;");
     createStatement.execute();
-    ptr = ShadowSQLiteConnection.nativeOpen(databasePath.getPath(), 0, "test connection", false, false);
+    ptr = ShadowSQLiteConnection.nativeOpen(databasePath.getPath(), 0, "test connection", false, false).longValue();
     CONNECTIONS = ReflectionHelpers.getStaticField(ShadowSQLiteConnection.class, "CONNECTIONS");
     conn = CONNECTIONS.getConnection(ptr);
   }
